fix(desktop): unbreak safe-url test on bun by splitting pure helpers (#3659)
`safe-url.ts` imports `shell` from electron at the module top level, so
bun's test runner can't load the file and the whole suite fails with
`SyntaxError: Export named 'shell' not found`. Moves the pure URL
helpers (`isSafeExternalUrl`, `externalUrlLogLabel`) to `scheme.ts`
and has the test import from there. `safeOpenExternal` stays in
`safe-url.ts` with the electron import; the barrel keeps the same
public surface.
