+++ apps/desktop/src/main/lib/safe-url/index.ts
@@ -1,5 +1,2 @@
-export {
-	externalUrlLogLabel,
-	isSafeExternalUrl,
-	safeOpenExternal,
-} from "./safe-url";
+export { safeOpenExternal } from "./safe-url";
+export { externalUrlLogLabel, isSafeExternalUrl } from "./scheme";
